Literals: move Case inside the companion
First, it is used only with Literals, and second, it avoids conflicts
with scala.meta.Case.
