+++ scalafmt-core/shared/src/main/scala/org/scalafmt/config/Literals.scala
@@ -1,5 +1,6 @@
 package org.scalafmt.config
 
+import Literals.Case
 import metaconfig._
 
 case class Literals(
@@ -16,4 +17,21 @@ object Literals {
   implicit val surface: generic.Surface[Literals] = generic.deriveSurface
   implicit val codec: ConfCodecEx[Literals] = generic.deriveCodecEx(Literals())
     .noTypos
+
+  sealed abstract class Case {
+    import Case._
+    def process(str: String): String = this match {
+      case Unchanged => str
+      case Lower => str.toLowerCase()
+      case Upper => str.toUpperCase()
+    }
+  }
+
+  object Case {
+    implicit val codec: ConfCodecEx[Case] = ReaderUtil
+      .oneOf[Case](Upper, Lower, Unchanged)
+    case object Upper extends Case
+    case object Lower extends Case
+    case object Unchanged extends Case
+  }
 }
